refactor: separate html to pdf conversion logic from the saving function, so that it is possible to retrieve the converted pdf result for further manipulation instead of saving directly
