+++ src/runtime/composables/exportToPDF.ts
@@ -1,24 +1,12 @@
-import jsPDF, { HTMLOptions, jsPDFOptions } from 'jspdf'
+import { HTMLOptions, jsPDFOptions } from 'jspdf'
+import { htmlToPdf } from './htmlToPdf'
 
 export const exportToPDF = async (
   fileName: string,
-  element?: HTMLElement,
+  element: HTMLElement,
   documentOptions?: jsPDFOptions,
-  options?: HTMLOptions
+  htmlOptions?: HTMLOptions
 ) => {
-  if (!(element instanceof HTMLElement)) {
-    throw new TypeError('usePDFExport: element is not a HTMLElement.')
-  }
-  const orientation = (element.offsetWidth > element.offsetHeight) ? 'l' : 'p'
-
-  // eslint-disable-next-line new-cap
-  const pdf = new jsPDF({
-    orientation: documentOptions?.orientation ?? orientation,
-    unit: documentOptions?.unit ?? 'px',
-    format: documentOptions?.format ?? 'A4',
-    encryption: documentOptions?.encryption
-  })
-
-  await pdf.html(element, options)
+  const pdf = await htmlToPdf(element, documentOptions, htmlOptions)
   return pdf.save(fileName)
 }
